+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 Unreleased in the current development version:
+- CLI script for automatic multiple experiment analysis submission (#1160, #1175)
+
+AQUA core complete list:
+- AQUA entry point for installation and catalog addition (#1131, #1134, #1146, #1168)
+
+## [v0.8.2]
+
+Main changes are: 
+1. `aqua-grids.yaml` file split in multiple files into `grids` folder
+2. Container for Levante
 
 AQUA core complete list:
 - Introduce a function to convert NetCDF to Zarr and zarr catalog entry for LRA (#1068)
@@ -481,7 +491,8 @@ This is mostly built on the `AQUA` `Reader` class which support for climate mode
 This is the AQUA pre-release to be sent to internal reviewers. 
 Documentations is completed and notebooks are working.
 
-[unreleased]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.8.1...HEAD
+[unreleased]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.8.2...HEAD
+[v0.8.2]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.8.1...v0.8.2
 [v0.8.1]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.8...v0.8.1
 [v0.8]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.7.3...v0.8
 [v0.7.3]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.7.2...v0.7.3